Refactor InvalidNamedMacroArgument to simplify and prepare for imports
My assumption that macro references are visited before their
definitions was faulty. This is due to the structure of the ModuleNode.

Fortunately, that node type allows us to simplify the logic, because
the ModuleNode can be queried for all defined macros.

I'll make a similar change to simplify BadArgumentCountInMacroCall next.